[ADD]CustomEdittext
