+++ PMS_Android/app/src/main/res/drawable/ic_baseline_done_24.xml
@@ -1,4 +1,5 @@
-<vector android:height="24dp" android:tint="#787478"
+<vector
+    android:height="24dp" android:tint="#787478"
     android:viewportHeight="24" android:viewportWidth="24"
     android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
     <path android:fillColor="@android:color/white" android:pathData="M9,16.2L4.8,12l-1.4,1.4L9,19 21,7l-1.4,-1.4L9,16.2z"/>
